flatten logic, removed inner
